requirements.txt
Django==4.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
dj-database-url==3.1.0
gunicorn==23.0.0
whitenoise==6.11.0
python-dotenv==1.2.1
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
psycopg2-binary==2.9.11
pillow==12.1.0
PyJWT==2.10.1
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3
requests==2.32.5
six==1.17.0
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
packaging==26.0
